# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Trivias y Preguntas",
        "description": "Documentación de la API para la aplicación de Trivias y Preguntas.",
        "contact": {
            "email": "test@email.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de Desarrollo"
        }
    ],
    "paths": {
        "/api/preguntas/crear": {
            "post": {
                "tags": [
                    "Preguntas"
                ],
                "summary": "Crear preguntas",
                "operationId": "505fe301ef8db824c714a4fcb8a2fe56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "required": [
                                        "enunciado",
                                        "opciones",
                                        "opcion_correcta_id",
                                        "nivel_dificultad"
                                    ],
                                    "properties": {
                                        "enunciado": {
                                            "type": "string",
                                            "example": "¿Qué es una entrevista de trabajo?"
                                        },
                                        "opciones": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "texto_opcion": {
                                                        "type": "string",
                                                        "example": "Una reunión social"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "opcion_correcta_id": {
                                            "type": "string",
                                            "example": "2"
                                        },
                                        "nivel_dificultad": {
                                            "type": "string",
                                            "enum": [
                                                "facil",
                                                "medio",
                                                "dificil"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Preguntas creadas"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
...
        "/api/respuestas/crear": {
            "post": {
                "tags": [
                    "Respuestas"
                ],
                "summary": "Registrar respuestas de una trivia",
                "operationId": "9152387a7fe3994fb925b6b5b228d07f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "jugador",
                                    "trivia_id",
                                    "preguntas"
                                ],
                                "properties": {
                                    "jugador": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jugador@email.com"
                                    },
                                    "trivia_id": {
                                        "type": "string",
                                        "example": "6658b60e9d63c84e8c0f0a11"
                                    },
                                    "preguntas": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "pregunta_id",
                                                "respuesta_jugador"
                                            ],
                                            "properties": {
                                                "pregunta_id": {
                                                    "type": "string",
                                                    "example": "664a23c48d5f6a1e9b7f8b4f"
                                                },
                                                "respuesta_jugador": {
                                                    "type": "string",
                                                    "example": "1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Respuestas guardadas correctamente"
                    },
                    "422": {
                        "description": "Validación fallida"
                    },
                    "404": {
                        "description": "Jugador no registrado"
                    },
                    "500": {
                        "description": "Error interno al guardar la respuesta"
                    }
                }
            }
        },
        "/api/respuestas/listar": {
            "get": {
                "tags": [
                    "Respuestas"
                ],
                "summary": "Listar respuestas de un jugador (por email y número de trivia opcional)",
                "operationId": "7b315725e78586f9d587dad430ffbdd3",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Correo electrónico del jugador",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "example": "jugador@email.com"
                        }
                    },
                    {
                        "name": "numero_trivia",
                        "in": "query",
                        "description": "Número de trivia (opcional)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de respuestas del jugador"
                    },
                    "422": {
                        "description": "Datos inválidos o incompletos"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa el token Bearer (ej. 'Bearer your_token_here')",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}